Expand `connections_signals` for multiple signals
Allow a connection to be simulated multiple signals as well as the
existing ability of a signal to simulate multiple connections. This is
necessary for building merge trees where a connection may need to be
simulated by multiple signals (o1 -> merge node ... -> o2).

Add a `all_signals` iterator to `Model` which yields each signal only
once.