Rewrite of filter functionality
This commit makes some major changes to the ways filters are handled in
C.  Please refer to the code for detailed comments.  In brief:

  - A filter is a function which accepts a dimensionality, an input
    vector, an output vector and some filter-specific state.
  - Instances of filters are instantiated with function pointers which
    point at the function which should be used to evaluate them

The way the data is packed to be transmitted to the C has changed to
represent these changes. Each filter entry (as written out by the
Python) contains at least four words. The first word indicates the
number of extra words that are the parameters for the filter; the second
word indexes an array of function pointers to find the initialisation
method for the given filter; the third word indicates the "width" of the
filter (the number of dimensions in the input and output vector) and the
fourth word contains flags. Additional words contain filter-specific
parameters (e.g., coefficients, indication of order, etc.).

An additional test has been included to check that the filters work
correctly in C. **No new filter types have been implemented in this
commit.**